+++ src/TrashLib/Services/CustomFormat/Models/CustomFormatData.cs
@@ -1,0 +1,11 @@
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+
+namespace TrashLib.Services.CustomFormat.Models;
+
+public record CustomFormatData(
+    string Name,
+    string TrashId,
+    int? Score,
+    [property: JsonExtensionData] JObject ExtraJson
+);