+++ src/TrashLib/Services/CustomFormat/Models/ProcessedConfigData.cs
@@ -1,0 +1,12 @@
+using TrashLib.Config.Services;
+
+namespace TrashLib.Services.CustomFormat.Models;
+
+public class ProcessedConfigData
+{
+    public ICollection<ProcessedCustomFormatData> CustomFormats { get; init; }
+        = new List<ProcessedCustomFormatData>();
+
+    public ICollection<QualityProfileConfig> QualityProfiles { get; init; }
+        = new List<QualityProfileConfig>();
+}