+++ src/TrashLib/Services/CustomFormat/Processors/GuideSteps/ICustomFormatStep.cs
@@ -1,0 +1,16 @@
+using TrashLib.Config.Services;
+using TrashLib.Services.CustomFormat.Models;
+using TrashLib.Services.CustomFormat.Models.Cache;
+
+namespace TrashLib.Services.CustomFormat.Processors.GuideSteps;
+
+public interface ICustomFormatStep
+{
+    IReadOnlyCollection<ProcessedCustomFormatData> ProcessedCustomFormats { get; }
+    IReadOnlyCollection<TrashIdMapping> DeletedCustomFormatsInCache { get; }
+    IReadOnlyCollection<(string, string)> CustomFormatsWithOutdatedNames { get; }
+    IDictionary<string, List<ProcessedCustomFormatData>> DuplicatedCustomFormats { get; }
+
+    void Process(IList<CustomFormatData> customFormatGuideData,
+        IReadOnlyCollection<CustomFormatConfig> config, CustomFormatCache? cache);
+}