+++ src/TrashLib/Services/Sonarr/ReleaseProfile/Guide/ISonarrGuideService.cs
@@ -1,8 +1,9 @@
+using TrashLib.Services.Common;
 using TrashLib.Services.Sonarr.QualityDefinition;
 
 namespace TrashLib.Services.Sonarr.ReleaseProfile.Guide;
 
-public interface ISonarrGuideService
+public interface ISonarrGuideService : IGuideService
 {
     IReadOnlyCollection<ReleaseProfileData> GetReleaseProfileData();
     ReleaseProfileData? GetUnfilteredProfileById(string trashId);